Implemented Room for trades

+++ app/src/main/java/com/example/stockgazer/ui/screens/chart/ChartViewModel.kt
@@ -7,15 +7,23 @@ import androidx.lifecycle.viewModelScope
 import com.example.stockgazer.data.datasource.AlpacaBarDatasource
 import com.example.stockgazer.data.datasource.AlpacaDetailsDatasource
 import com.example.stockgazer.data.entities.FavoriteStock
+import com.example.stockgazer.data.entities.Trade
 import com.example.stockgazer.data.response.SnapshotResponse
 import com.example.stockgazer.storage.StockGazerDatabase
 import com.example.stockgazer.util.ResourceZoneIdProvider
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.flatMapLatest
+import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import java.time.LocalDate
+import java.time.LocalDateTime
 import java.time.LocalTime
 import javax.inject.Inject
 
@@ -25,10 +33,8 @@ class ChartViewModel @Inject constructor(
     private val alpacaDetailsDatasource: AlpacaDetailsDatasource,
     val zoneIdProvider: ResourceZoneIdProvider,
     @ApplicationContext private val context: Context,
-    ) : ViewModel() {
+) : ViewModel() {
     private val database = StockGazerDatabase.getDatabase(context)
-
-    val isFavorite = database.favoriteStockDao().isStockFavorite("MSFT").asFlow()
 
     private var _showTradeCreationModal = MutableStateFlow(false)
     val showTradeCreationModal = _showTradeCreationModal.asStateFlow()
@@ -36,10 +42,20 @@ class ChartViewModel @Inject constructor(
     private var _companyInfo = MutableStateFlow(CompanyInfo("", ""))
     val companyInfo = _companyInfo.asStateFlow()
 
+    @OptIn(ExperimentalCoroutinesApi::class)
+    val isFavorite = _companyInfo
+        .filterNotNull()
+        .flatMapLatest { companyInfo ->
+            database.favoriteStockDao().isStockFavorite(companyInfo.symbol).asFlow()
+        }.stateIn(
+            scope = viewModelScope,
+            started = SharingStarted.WhileSubscribed(5000),
+            initialValue = false
+        )
+
     private var _latestPrice = MutableStateFlow(
         LatestPrice(
-            value = 0.0,
-            dailyPercentChange = 0.0
+            value = 0.0, dailyPercentChange = 0.0
         )
     )
     val latestPrice = _latestPrice.asStateFlow()
@@ -47,15 +63,23 @@ class ChartViewModel @Inject constructor(
     private var _bars = MutableStateFlow(BarPeriod())
     val bars = _bars.asStateFlow()
 
-    private val defaultTrade = Trade(
+    private val defaultTradeFormData = TradeFormData(
         date = LocalDate.now(zoneIdProvider.getTimeZone()),
         time = LocalTime.now(zoneIdProvider.getTimeZone()),
     )
-    private var _currentTrade = MutableStateFlow(defaultTrade)
+    private var _currentTrade = MutableStateFlow(defaultTradeFormData)
     val currentTrade = _currentTrade.asStateFlow()
 
-    private var _trades = MutableStateFlow(emptyList<Trade>())
-    val trades = _trades.asStateFlow()
+    @OptIn(ExperimentalCoroutinesApi::class)
+    val trades: StateFlow<List<Trade>> = _companyInfo
+        .filterNotNull()
+        .flatMapLatest { companyInfo ->
+            database.tradeDao().getAllTradesOfSymbol(companyInfo.symbol).asFlow()
+        }.stateIn(
+            scope = viewModelScope,
+            started = SharingStarted.WhileSubscribed(5000),
+            initialValue = emptyList()
+        )
 
     private var _chartLoadState = MutableStateFlow(ChartLoadState())
     val loadState = _chartLoadState.asStateFlow()
@@ -67,57 +91,45 @@ class ChartViewModel @Inject constructor(
     }
 
     private fun loadDetails(symbol: String) {
-        alpacaDetailsDatasource.getStockOverview(symbol,
-            onSuccess = {
-                viewModelScope.launch {
-                    _companyInfo.emit(CompanyInfo(it.name, symbol))
-                    _chartLoadState.emit(_chartLoadState.value.copy(detailsLoaded = true))
-                }
-            },
-            onFail = {},
-            loadingFinished = {}
-        )
+        alpacaDetailsDatasource.getStockOverview(symbol, onSuccess = {
+            viewModelScope.launch {
+                _companyInfo.emit(CompanyInfo(it.name, symbol))
+                _chartLoadState.emit(_chartLoadState.value.copy(detailsLoaded = true))
+            }
+        }, onFail = {}, loadingFinished = {})
     }
 
     private fun loadSnapshot(symbol: String) {
-        alpacaBarDatasource.getSnapshotFromSymbols(listOf(symbol),
-            onSuccess = {
-                val symbolsSnapshot: SnapshotResponse = it[symbol]!!
+        alpacaBarDatasource.getSnapshotFromSymbols(listOf(symbol), onSuccess = {
+            val symbolsSnapshot: SnapshotResponse = it[symbol]!!
 
-                viewModelScope.launch {
-                    val latestPriceFetched = LatestPrice.fromSnapshotResponse(symbolsSnapshot)
-                    if (!_latestPrice.value.isLoaded()) {
-                        _currentTrade.emit(_currentTrade.value.copy(price = latestPriceFetched.value.toString()))
-                    }
-                    _latestPrice.emit(latestPriceFetched)
-
-                    _chartLoadState.emit(_chartLoadState.value.copy(snapshotLoaded = true))
+            viewModelScope.launch {
+                val latestPriceFetched = LatestPrice.fromSnapshotResponse(symbolsSnapshot)
+                if (!_latestPrice.value.isLoaded()) {
+                    _currentTrade.emit(_currentTrade.value.copy(price = latestPriceFetched.value.toString()))
                 }
-            }, onFail = {},
-            loadingFinished = {}
-        )
+                _latestPrice.emit(latestPriceFetched)
+
+                _chartLoadState.emit(_chartLoadState.value.copy(snapshotLoaded = true))
+            }
+        }, onFail = {}, loadingFinished = {})
     }
 
     private fun loadBars(symbol: String) {
-        alpacaBarDatasource.getBarsFromSymbol(symbol,
-            onSuccess = {
-                viewModelScope.launch {
-                    _bars.emit(BarPeriod.fromBarResponse(it))
-                    _chartLoadState.emit(_chartLoadState.value.copy(barsLoaded = true))
-                }
-            },
-            onFail = {},
-            loadingFinished = {})
+        alpacaBarDatasource.getBarsFromSymbol(symbol, onSuccess = {
+            viewModelScope.launch {
+                _bars.emit(BarPeriod.fromBarResponse(it))
+                _chartLoadState.emit(_chartLoadState.value.copy(barsLoaded = true))
+            }
+        }, onFail = {}, loadingFinished = {})
     }
 
     fun toggleFavorite(isFavorite: Boolean) {
         viewModelScope.launch {
             val symbol = _companyInfo.value.symbol
 
-            if (isFavorite)
-                database.favoriteStockDao().deleteFavoriteStock(symbol)
-            else
-                database.favoriteStockDao().insert(FavoriteStock(symbol))
+            if (isFavorite) database.favoriteStockDao().deleteFavoriteStock(symbol)
+            else database.favoriteStockDao().insert(FavoriteStock(symbol))
         }
     }
 
@@ -170,11 +182,22 @@ class ChartViewModel @Inject constructor(
 
     fun submitTrade() {
         viewModelScope.launch {
-            _trades.emit(_trades.value + _currentTrade.value)
+            val trade = _currentTrade.value
+            database.tradeDao().insert(
+                Trade(
+                    _companyInfo.value.symbol,
+                    trade.type,
+                    trade.amount,
+                    trade.price,
+                    LocalDateTime.of(trade.date, trade.time)
+                )
+            )
+
             _currentTrade.emit(
-                defaultTrade.copy(price = _latestPrice.value.value.toString())
+                defaultTradeFormData.copy(price = _latestPrice.value.value.toString())
             )
             _showTradeCreationModal.emit(false)
         }
     }
 }
+
